Fix bf16 FFT crash in the loss on MPS (and compute loss in fp32)
Training on Apple MPS failed with "RuntimeError: Unsupported dtype BFloat16" in
spectral_harmonic_loss: under autocast the loss received bf16 tensors and called
torch.fft.rfft on them, which MPS rejects. (CPU tolerated it, so the sandbox
tests missed it — same root cause as the earlier FNO fix, but in the loss.)

Fixes:
- spectral_harmonic_loss now runs the FFT in float32 with autocast disabled.
- Trainer computes the whole geometric loss in float32 outside autocast (the
  standard mixed-precision pattern; geodesic arccos is also bf16-imprecise).
- Regression tests feed bf16 tensors through the spectral and composite losses.

Verified: scripts/train.py runs end to end (forward under bf16 autocast, loss in
fp32) and saves checkpoints. 93 tests pass.

+++ src/kalachakra/training/trainer.py
@@ -104,10 +104,13 @@ class Trainer:
         self.optimizer.zero_grad(set_to_none=True)
         with torch.autocast(device_type=self.device.type, dtype=self._amp_dtype):
             recon, _z = self.model(e)
-            # Reshape channel axis back to (..., B, 5) for the geometric loss.
-            recon_f = recon.unflatten(-1, (-1, 5))
-            target_f = e.unflatten(-1, (-1, 5))
-            total, parts = self.criterion(recon_f, target_f)
+
+        # Compute the loss in float32, outside autocast: the geometric loss uses
+        # FFTs and arccos, which are unsupported or imprecise in bf16/fp16 (and
+        # error on MPS). This is the standard mixed-precision pattern.
+        recon_f = recon.float().unflatten(-1, (-1, 5))
+        target_f = e.float().unflatten(-1, (-1, 5))
+        total, parts = self.criterion(recon_f, target_f)
 
         total.backward()
         if self.cfg.grad_clip:
